requirements.txt
django==3.0.5
channels==2.3.1
psycopg2==2.8.4
python-dateutil==2.8.1
celery==4.3.0
numpy==1.17.3
pathfinding==0.0.4
django-extensions==2.1.7
channels-redis==2.4.2
djangorestframework==3.11.0
django-cors-headers==3.2.1